scripts/requirements.txt
jinja2==3.0.1
pyyaml==5.4.1
